Fixes #103
This moves whole logics of channel work to single file 'Channels', which
brings all the channel state to the singletone, which has needed
clearing logics.
